Improve output
